Use updated type for cpu (uint) and memory (uint)
This will reduce some of interger overflow checks as part of golint ci
run.

+++ pkg/crc/machine/config/config.go
@@ -8,9 +8,9 @@ type MachineConfig struct {
 
 	// Virtual machine configuration
 	Name              string
-	Memory            int
-	CPUs              int
-	DiskSize          int
+	Memory            uint
+	CPUs              uint
+	DiskSize          uint64
 	ImageSourcePath   string
 	ImageFormat       string
 	SSHKeyPath        string
